Architectural Refactoring: [1] Model package organization will not separate Steps, DTOs and Pages in classes; [2] Actually have now a package for each regarding Page, which java file is now within its proper classes (DTO and Step). [3] Abstracts or repetitive DTO will be placed in proper packages.
