+++ src/test/java/model/pages/AbstractPage.java
@@ -12,7 +12,7 @@ import core.context.TestContext;
 import core.utils.webutils.JavaScriptUtils;
 import core.utils.webutils.Waits;
 
-public abstract class Page {
+public abstract class AbstractPage {
 
     protected WebDriver driver;
     protected FluentWait<WebDriver> isElemClickable;
@@ -20,7 +20,7 @@ public abstract class Page {
     protected JavaScriptUtils javaScriptUtils;
     protected Waits waits;
 
-    public Page(WebDriver wdriver) {
+    public AbstractPage(WebDriver wdriver) {
 	driver = wdriver;
 	this.javaScriptUtils = new JavaScriptUtils(driver);
 	this.elemIsPresent = new FluentWait<WebDriver>(driver);
