+++ src/test/java/model/utils/ExcelUtils.java
@@ -4,11 +4,13 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
+import org.apache.poi.ss.usermodel.CellType;
 import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.xssf.usermodel.XSSFCell;
 import org.apache.poi.xssf.usermodel.XSSFRow;
 import org.apache.poi.xssf.usermodel.XSSFSheet;
 
-import model.dtos.UserFormDTO;
+import model.formsubmit.UserFormDTO;
 
 public class ExcelUtils {
 
@@ -30,5 +32,23 @@ public class ExcelUtils {
 	}
 	return userFormList;
 
+    }
+
+    public String getStringCellValue(XSSFCell cell) {
+	switch (cell.getCellType()) {
+	case STRING:
+	case FORMULA:
+	    return cell.getStringCellValue();
+	case BOOLEAN:
+	    return String.valueOf(cell.getBooleanCellValue());
+	case ERROR:
+	    return String.valueOf(cell.getErrorCellValue());
+	case NUMERIC:
+	    return String.valueOf(cell.getNumericCellValue());
+	default: {
+	    cell.setCellType(CellType.STRING);
+	    return getStringCellValue(cell);
+	}
+	}
     }
 }
