+++ src/test/java/runners/RunnerTest.java
@@ -10,7 +10,7 @@ import steps.SetupStep;
 
 @CucumberOptions(
 	features = { "src/test/resources/features/form-submit.feature" },
-	glue = { "steps" , "hooks" },
+	glue = { "model.formsubmit" , "hooks" },
 	plugin = {"summary", "pretty" },
 	tags = {"@ID_0001 or @ID_0002 or @ID_0003 or @ID_0004 or @ID_0005 or @ID_0006"},
 	snippets = SnippetType.CAMELCASE,
